requirements.txt
# Web 框架和工具
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
aiofiles==24.1.0
websockets==15.0.1

# 数据库和 ORM
sqlalchemy==2.0.28
alembic==1.12.1

# 数据验证和配置
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1

# AI 和机器学习
sentence-transformers==2.5.1
faiss-cpu==1.10.0
torch==2.2.1
transformers==4.38.2

# PDF 和文档处理
pypdf2==3.0.1
python-docx==1.1.0
pymupdf==1.24.0
pdfplumber==0.11.6
pdf2image==1.17.0

# 图像处理
Pillow==10.2.0
opencv-python==4.9.0.80
pytesseract==0.3.10
poppler-utils

# 数据处理和分析
pandas==2.2.1
numpy==1.26.4

# HTTP 和网络
requests==2.31.0

# 工具和实用程序
tqdm==4.66.2
six==1.16.0
python-dateutil==2.9.0.post0
pytz==2025.2
